requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
PyJWT==2.3.0
pytz==2022.1
PyYAML==6.0
quickchart.io==1.0.1
requests==2.27.1
slack-bolt==1.13.0
slack-sdk==3.15.2
urllib3==1.26.9
Flask>=1.1
gunicorn>=20